+++ docs/CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - multiple modal windows might be open at a time
 - control-click to move between belt and inventory
 - added 'target mode' to select items or objects exclusively
+- support swapping of hotkey-sets
 ### Gameplay
 - SAVE-files are NOT backwards compatible
 - altered abilities (hellfire)